Removed Tree animator & fixxed player run animation
Removing Tree animator saved a lot of performance due to the amount of trees there are and it fixxed that the tree can now shrink through code because the animator was overriding it.

Then I made the player animation be inside sub states and be played from any state, it wasn't really flexible before.